# openapi.yaml
openapi: 3.1.0
info:
  title: DevMate Core API Specification
  description: |
    Production-ready REST and RPC API contract for the DevMate Telegram-based Personal Operating System platform.
    Defines schemas, endpoints, error response wrappers, webhooks, and callbacks for delivery adapters and future clients.
  version: 1.0.0
  termsOfService: https://devmate.internal/terms
  contact:
    name: DevMate API Platform Team
    email: platform-api@devmate.internal
    url: https://devmate.internal/support
  license:
    name: Proprietary License
    url: https://devmate.internal/license
...
servers:
  - url: https://api.devmate.internal/v1
    description: Production Server Gateway
  - url: https://staging-api.devmate.internal/v1
    description: Staging Environment Server
  - url: https://dev-api.devmate.internal/v1
    description: Development Integration Server
  - url: http://localhost:8080/v1
    description: Local Developer Container Server
security:
  - BearerAuth: []
paths:
...
  /reminders:
    get:
      summary: List reminders
      description: Fetches a list of active scheduler reminders.
      operationId: listReminders
      tags:
        - Reminders
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Reminders retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create reminder alert
      description: Registers a new alert schedule.
      operationId: createReminder
      tags:
        - Reminders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /reminders/{id}:
    get:
      summary: Get reminder details
      description: Retrieves a single reminder schedule configuration.
      operationId: getReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reminder details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update reminder trigger details
      description: Modifies text, recurrence rules, or target trigger timestamps.
      operationId: updateReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder rescheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Remove reminder alert
      description: Soft deletes a reminder schedule.
      operationId: deleteReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Reminder removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /reminders/{id}/snooze:
    post:
      summary: Snooze reminder trigger
      description: Delays scheduled trigger execution time.
      operationId: snoozeReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - snoozeMinutes
              properties:
                snoozeMinutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  example: 15
      responses:
        '200':
          description: Reminder trigger rescheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /reminders/{id}/resume:
    post:
      summary: Resume reminder triggers
      description: Reactivates paused scheduler alerts.
      operationId: resumeReminder
      tags:
        - Reminders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reminder trigger resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived session access token passed in Authorization headers.
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Secure HTTP-only refresh token used to request new access tokens.
    TelegramSecretVerification:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Shared bot secret payload validating inbound Webhook calls.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Reserved API key token for external script integrations.
    OAuth2Placeholder:
      type: oauth2
      description: Reserved for future oauth integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.devmate.internal/oauth/authorize
          tokenUrl: https://auth.devmate.internal/oauth/token
          scopes:
            read: Grant view permissions on owned resources.
            write: Grant modification permissions on owned resources.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Max record items count.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      description: Encoded cursor pointer for page offsets.
      required: false
      schema:
        type: string
...
  headers:
    CorrelationIdResponseHeader:
      description: Ingress trace correlation tracker ID UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Input validation constraints violations.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationIdResponseHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
...
    Unauthorized:
      description: Authentication failure due to missing or expired tokens.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServerError:
      description: Internal system error log tracker.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    CreateReminderRequest:
      type: object
      required:
        - text
        - triggerTime
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        triggerTime:
          type: string
          format: date-time
        recurrence:
          type: string
    UpdateReminderRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        triggerTime:
          type: string
          format: date-time
        recurrence:
          type: string
    ReminderResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ReminderDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ReminderDetail:
      type: object
      required:
        - id
        - text
        - triggerTime
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        triggerTime:
          type: string
          format: date-time
        recurrence:
          type: string
        status:
          $ref: '#/components/schemas/ReminderStatus'
        version:
          type: integer
    ReminderListResponse:
      type: object
      required:
        - success
        - data
        - pagination
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReminderDetail'
        pagination:
          $ref: '#/components/schemas/CursorPaginationMeta'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ResponseMeta:
      type: object
      required:
        - correlationId
        - traceId
      properties:
        correlationId:
          type: string
          format: uuid
        traceId:
          type: string
          format: uuid
    CursorPaginationMeta:
      type: object
      required:
        - hasMore
      properties:
        nextCursor:
          type: string
        hasMore:
          type: boolean
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Request resource not found.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ValidationErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - details
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message:
              type: string
              example: Invalid input parameters.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: object
                required:
                  - field
                  - issue
                properties:
                  field:
                    type: string
                    example: amount
                  issue:
                    type: string
                    example: Value must be a positive string.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ReminderStatus:
      type: string
      enum: [PENDING, FIRED, MISSED, SNOOZED, PAUSED, COMPLETED]
...